refactor: create utils for auth
